Fixed normalsShader, Fixed NPC173, Added HW skinning to zBuffer shader, Restored crouching
Now normalsShader gets tangent and binormal from attributes instead of
creating them with the normal. Made to work with deferred lighting.

NPC173 broke because of the scaling change from 0.75 to 0.1. Now it
works as before, but it's still a bit buggy (should probably replace
occlusion query with something that can be performed without rendering).

Readded crouching, had to modify the kinematic player controller to stop
breaking with size changes.